+++ devices/dev_win.go
@@ -1,4 +1,4 @@
-package device
+package devices
 
 import (
 	"os"
